separate vtkjs serialization from pane implementation and add set_serializer method
